refactor: comprehensive platform overhaul (phases 1-5)
Phase 1 - Critical Bug Fixes:
- Fix floating-point fee calculation to integer arithmetic (TransferService, WalletService)
- Fix UserId type mismatch in TransactionLedger (int -> uint)
- Restrict RecalculateBalance to admin-only
- Add missing database indexes (validation_case, consultation_request, validation_case_log)
- Add fail-fast config validation for FEATURE_SERVICE_URL
- Add daily withdrawal limits

Phase 2 - Backend Reliability:
- Fix N+1 query in matching score with batch loading
- Add distributed locking for SLA worker
- Migrate rate limiting to Redis-based sliding window
- Add request correlation IDs (Go + .NET)
- Add circuit breaker for inter-service communication
- Add PQC key lookup caching

Phase 3 - Frontend Modernization:
- Add error boundaries for account, validation-cases, user, threads
- Add page metadata via layouts (account, security)
- Split PasskeySettings -> PasskeyList + PasskeySettings
- Split TOTPSettings -> TOTPSetupWizard + TOTPDisableForm + TOTPSettings
- Add register server action
- Add optimistic UI for consultation approve/reject
- Add streaming SSR with Suspense for validation-cases listing
- Rename .js components to .jsx
- Add modern CSS enhancements (container queries, view transitions)
- Integrate logger with Sentry for production
- Unify admin auth with main JWT (bridge pattern)

Phase 4 - Testing:
- Add 74 new unit tests across 5 test suites (format, api, adminAuth, tokenRefresh, logger)
- Total: 130 tests, 9 suites, all passing

Phase 5 - Observability:
- Add frontend-test job to CI pipeline
- Security headers already configured (Permissions-Policy, X-Frame-Options, etc.)
- Sentry fully integrated with session replay

+++ feature-service/src/FeatureService.Api/Services/WalletService.cs
@@ -285,7 +285,7 @@ public class WalletService : IWalletService
     public async Task<long> RecalculateBalanceFromLedgerAsync(uint userId)
     {
         var filter = Builders<TransactionLedger>.Filter.And(
-            Builders<TransactionLedger>.Filter.Eq(l => l.UserId, (int)userId),
+            Builders<TransactionLedger>.Filter.Eq(l => l.UserId, userId),
             Builders<TransactionLedger>.Filter.Eq(l => l.Status, TransactionStatus.Completed)
         );
 
@@ -700,7 +700,7 @@ public class WalletService : IWalletService
 
             var ledgerEntry = new TransactionLedger
             {
-                UserId = (int)userId,
+                UserId = userId,
                 EntryType = entryType,
                 Amount = amount,
                 BalanceAfter = balanceAfter,
